DEV: use outlet to mount new UI
Outlet was introduced in this PR https://github.com/discourse/discourse/pull/22897
