+++ javascripts/discourse/connectors/below-filter-input/filter-options.hbs
@@ -1,0 +1,1 @@
+<Navigation::FilterOptions @newQueryString={{this.args.newQueryString}} />
